+++ packages/paraswap/src/Paraswap.test.ts
@@ -1,23 +1,40 @@
 import { GreaterThanOrEqual, apply } from '@rabbitholegg/questdk/filter'
+import { ActionType } from '@rabbitholegg/questdk'
 import { describe, expect, test } from 'vitest'
-import { SWAP_MULTI, SWAP_SIMPLE } from './test-transactions'
-import { swap } from './Paraswap.js'
-import { ARB_ONE_CHAIN_ID } from './chain-ids.js'
-import { parseEther, type Address } from 'viem'
-import { PARASWAP_ABI } from './abi.js'
-const USDT_ADDRESS = '0xFd086bC7CD5C481DCC9C85ebE478A1C0b69FCbb9'
-const USDCE_ADDRESS = '0xFF970A61A04b1cA14834A43f5dE4533eBDDB5CC8'
-const VELA_ADDRESS = '0x088cd8f5ef3652623c22d48b1605dcfe860cd704'
-const AUGUSTUS_SWAPPER_ARBITRUM = '0xdef171fe48cf0115b1d80b88dc8eab59176fee57'
+import {
+  MULTI_DEPOSIT,
+  SIMPLE_DEPOSIT,
+  PROD_SWAP_SIMPLE,
+  SIMPLE_SWAP_ZYBER_TO_ETH,
+  SINGLE_SWAP_EURO,
+  SINGLE_SWAP_USDC_BTC,
+  SWAP_MEGA,
+  SWAP_MULTI,
+  SWAP_SIMPLE,
+  SWAP_BALANCER,
+  SWAP_CURVE,
+  UNISWAP_V3_SWAP,
+  WETH_PROD_TEST,
+} from './test-transactions'
+import { stake, swap, getSupportedTokenAddresses } from './Paraswap.js'
+import { ARB_ONE_CHAIN_ID, OPTIMISM_CHAIN_ID } from './chain-ids.js'
+import { Tokens } from './utils'
+import { parseEther, parseUnits } from 'viem'
+import { PARASWAP_SWAP_ABI } from './abi.js'
+import type {
+  FilterObject,
+  TransactionFilter,
+} from '@rabbitholegg/questdk/dist/types/filter/types'
+const AUGUSTUS_SWAPPER_ARBITRUM = '0xDEF171Fe48CF0115B1d80b88dc8eAB59176FEe57'
 
 describe('Given the paraswap plugin', () => {
-  describe('When handling the bridge', () => {
+  describe('When handling the swap', () => {
     test('should return a valid action filter', async () => {
       const filter = await swap({
         chainId: ARB_ONE_CHAIN_ID,
         contractAddress: AUGUSTUS_SWAPPER_ARBITRUM,
-        tokenIn: USDT_ADDRESS.toLowerCase() as Address,
-        tokenOut: USDCE_ADDRESS.toLowerCase() as Address,
+        tokenIn: Tokens.USDT,
+        tokenOut: Tokens.USDCE,
         amountIn: GreaterThanOrEqual(339000000),
       })
 
@@ -25,69 +42,61 @@ describe('Given the paraswap plugin', () => {
         chainId: ARB_ONE_CHAIN_ID,
         to: AUGUSTUS_SWAPPER_ARBITRUM,
         input: {
-          $abiAbstract: PARASWAP_ABI,
+          $abi: PARASWAP_SWAP_ABI,
           $or: [
             {
-              assets: [
-                '0xfd086bc7cd5c481dcc9c85ebe478a1c0b69fcbb9',
-                '0xff970a61a04b1ca14834a43f5de4533ebddb5cc8',
-              ],
-              fromAmount: {
-                $gte: '339000000',
-              },
-              funds: {},
-            },
-            {
-              assets: [
-                '0xfd086bc7cd5c481dcc9c85ebe478a1c0b69fcbb9',
-                '0xff970a61a04b1ca14834a43f5de4533ebddb5cc8',
-              ],
-              fromAmount: {
-                $gte: '339000000',
-              },
-              funds: {},
-            },
-            {
-              fromAmount: {
-                $gte: '339000000',
-              },
-              fromToken: '0xfd086bc7cd5c481dcc9c85ebe478a1c0b69fcbb9',
-              toToken: '0xff970a61a04b1ca14834a43f5de4533ebddb5cc8',
-            },
-            {
-              fromAmount: {
-                $gte: '339000000',
-              },
-              fromToken: '0xfd086bc7cd5c481dcc9c85ebe478a1c0b69fcbb9',
-              toToken: '0xff970a61a04b1ca14834a43f5de4533ebddb5cc8',
-            },
-            {
-              params: {
-                amountIn: {
+              data: {
+                fromToken: '0xFd086bC7CD5C481DCC9C85ebE478A1C0b69FCbb9',
+                fromAmount: {
                   $gte: '339000000',
                 },
-                tokenIn: '0xfd086bc7cd5c481dcc9c85ebe478a1c0b69fcbb9',
-                tokenOut: '0xff970a61a04b1ca14834a43f5de4533ebddb5cc8',
-              },
-            },
-            {
-              params: {
-                amountIn: {
-                  $gte: '339000000',
-                },
-                path: '0xfd086bc7cd5c481dcc9c85ebe478a1c0b69fcbb9ff970a61a04b1ca14834a43f5de4533ebddb5cc8',
+                toToken: '0xFF970A61A04b1cA14834A43f5dE4533eBDDB5CC8',
               },
             },
             {
               data: {
+                fromToken: '0xFd086bC7CD5C481DCC9C85ebE478A1C0b69FCbb9',
                 fromAmount: {
                   $gte: '339000000',
                 },
-                fromToken: '0xfd086bc7cd5c481dcc9c85ebe478a1c0b69fcbb9',
                 path: {
                   $last: {
-                    to: '0xff970a61a04b1ca14834a43f5de4533ebddb5cc8',
+                    to: '0xFF970A61A04b1cA14834A43f5dE4533eBDDB5CC8',
                   },
+                },
+              },
+            },
+            {
+              data: {
+                fromToken: '0xFd086bC7CD5C481DCC9C85ebE478A1C0b69FCbb9',
+                fromAmount: {
+                  $gte: '339000000',
+                },
+                path: {
+                  $last: {
+                    path: {
+                      $last: {
+                        to: '0xFF970A61A04b1cA14834A43f5dE4533eBDDB5CC8',
+                      },
+                    },
+                  },
+                },
+              },
+            },
+            {
+              data: {
+                assets: {
+                  $and: [
+                    {
+                      $first: '0xFd086bC7CD5C481DCC9C85ebE478A1C0b69FCbb9',
+                    },
+                    {
+                      $last: '0xFF970A61A04b1cA14834A43f5dE4533eBDDB5CC8',
+                    },
+                  ],
+                },
+                fromAmount: {
+                  $gte: '339000000',
                 },
               },
             },
@@ -101,30 +110,140 @@ describe('Given the paraswap plugin', () => {
       const filter = await swap({
         chainId: ARB_ONE_CHAIN_ID,
         contractAddress: AUGUSTUS_SWAPPER_ARBITRUM,
-        tokenIn: USDT_ADDRESS.toLowerCase() as Address,
-        tokenOut: USDCE_ADDRESS.toLowerCase() as Address,
+        tokenIn: Tokens.USDT,
+        tokenOut: Tokens.USDCE,
         amountIn: GreaterThanOrEqual(339000000),
       })
       expect(apply(transaction, filter)).to.be.true
+    })
+    test('should pass filter with valid production simple transactions', async () => {
+      const transaction = PROD_SWAP_SIMPLE
+      const testFilter: TransactionFilter = {
+        to: AUGUSTUS_SWAPPER_ARBITRUM,
+        input: {
+          //@ts-ignore
+          $abi: PARASWAP_SWAP_ABI,
+          $or: [
+            {
+              data: {
+                assets: {
+                  $and: [
+                    {
+                      $first: '0x82aF49447D8a07e3bd95BD0d56f35241523fBab1',
+                    },
+                    {
+                      $last: '0x0000000000000000000000000000000000000000',
+                    },
+                  ],
+                },
+              },
+            } as FilterObject,
+            {
+              data: {
+                path: {
+                  $last: {
+                    path: {
+                      $last: {
+                        to: '0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE',
+                      },
+                    },
+                  },
+                },
+                fromToken: '0x82aF49447D8a07e3bd95BD0d56f35241523fBab1',
+              },
+            } as FilterObject,
+            {
+              data: {
+                path: {
+                  $last: {
+                    to: '0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE',
+                  },
+                },
+                fromToken: '0x82aF49447D8a07e3bd95BD0d56f35241523fBab1',
+              },
+            } as FilterObject,
+            {
+              data: {
+                toToken: '0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE',
+                fromToken: '0x82aF49447D8a07e3bd95BD0d56f35241523fBab1',
+              },
+            } as FilterObject,
+          ],
+        },
+        chainId: 42161,
+      }
+      expect(apply(transaction, testFilter)).to.be.true
     })
     test('should not pass filter with invalid simple transactions', async () => {
       const transaction = SWAP_SIMPLE
       const filter = await swap({
         chainId: ARB_ONE_CHAIN_ID,
         contractAddress: AUGUSTUS_SWAPPER_ARBITRUM,
-        tokenIn: USDT_ADDRESS.toLowerCase() as Address,
-        tokenOut: USDCE_ADDRESS.toLowerCase() as Address,
+        tokenIn: Tokens.USDT,
+        tokenOut: Tokens.USDCE,
         amountIn: GreaterThanOrEqual(339000000000),
       })
       expect(apply(transaction, filter)).to.be.false
     })
+
+    test('should pass filter with valid singe swap WETH -> USDC', async () => {
+      const transaction = WETH_PROD_TEST
+      const filter = await swap({
+        chainId: ARB_ONE_CHAIN_ID,
+        contractAddress: AUGUSTUS_SWAPPER_ARBITRUM,
+        tokenIn: Tokens.WETH,
+        tokenOut: Tokens.USDCE,
+      })
+      expect(apply(transaction, filter)).to.be.true
+    })
+
+    test('should pass filter with valid singe swap USDC -> WBTC', async () => {
+      const transaction = SINGLE_SWAP_USDC_BTC
+      const filter = await swap({
+        chainId: ARB_ONE_CHAIN_ID,
+        contractAddress: AUGUSTUS_SWAPPER_ARBITRUM,
+        tokenIn: Tokens.USDCE,
+      })
+      expect(apply(transaction, filter)).to.be.true
+    })
+
+    test('should pass filter with valid singe swap USDC -> EURO', async () => {
+      const transaction = SINGLE_SWAP_EURO
+      const filter = await swap({
+        chainId: ARB_ONE_CHAIN_ID,
+        contractAddress: AUGUSTUS_SWAPPER_ARBITRUM,
+        tokenIn: Tokens.USDCE,
+      })
+      expect(apply(transaction, filter)).to.be.true
+    })
+
+    test('should pass filter with valid Uniswap swap', async () => {
+      const transaction = UNISWAP_V3_SWAP
+      const filter = await swap({
+        chainId: ARB_ONE_CHAIN_ID,
+        contractAddress: AUGUSTUS_SWAPPER_ARBITRUM,
+        tokenIn: Tokens.ETH,
+      })
+      expect(apply(transaction, filter)).to.be.true
+    })
+
+    test('should pass filter with valid Simple swap Zyber -> ETH', async () => {
+      const transaction = SIMPLE_SWAP_ZYBER_TO_ETH
+      const filter = await swap({
+        chainId: ARB_ONE_CHAIN_ID,
+        contractAddress: AUGUSTUS_SWAPPER_ARBITRUM,
+        tokenOut: Tokens.ETH,
+      })
+      expect(apply(transaction, filter)).to.be.true
+    })
+
     test('should pass filter with valid multi transactions', async () => {
       const transaction = SWAP_MULTI
       const filter = await swap({
         chainId: ARB_ONE_CHAIN_ID,
         contractAddress: AUGUSTUS_SWAPPER_ARBITRUM,
-        tokenIn: USDCE_ADDRESS.toLowerCase() as Address,
-        tokenOut: VELA_ADDRESS.toLowerCase() as Address,
+        tokenIn: Tokens.USDCE,
+        tokenOut: Tokens.VELA,
         amountOut: GreaterThanOrEqual(parseEther('0.037')),
       })
       expect(apply(transaction, filter)).to.be.true
@@ -134,11 +253,122 @@ describe('Given the paraswap plugin', () => {
       const filter = await swap({
         chainId: ARB_ONE_CHAIN_ID,
         contractAddress: AUGUSTUS_SWAPPER_ARBITRUM,
-        tokenIn: USDT_ADDRESS.toLowerCase() as Address,
-        tokenOut: VELA_ADDRESS.toLowerCase() as Address,
+        tokenIn: Tokens.USDT,
+        tokenOut: Tokens.VELA,
         amountOut: GreaterThanOrEqual(parseEther('0.037')),
       })
       expect(apply(transaction, filter)).to.be.false
+    })
+    test('should pass filter with valid simple transactions', async () => {
+      const transaction = SWAP_SIMPLE
+      const filter = await swap({
+        chainId: ARB_ONE_CHAIN_ID,
+        contractAddress: AUGUSTUS_SWAPPER_ARBITRUM,
+        tokenIn: Tokens.USDT,
+        tokenOut: Tokens.USDCE,
+        amountIn: GreaterThanOrEqual(339000000),
+      })
+      expect(apply(transaction, filter)).to.be.true
+    })
+    test('should not pass filter with invalid simple transactions', async () => {
+      const transaction = SWAP_SIMPLE
+      const filter = await swap({
+        chainId: ARB_ONE_CHAIN_ID,
+        contractAddress: AUGUSTUS_SWAPPER_ARBITRUM,
+        tokenIn: Tokens.USDT,
+        tokenOut: Tokens.USDCE,
+        amountIn: GreaterThanOrEqual(339000000000),
+      })
+      expect(apply(transaction, filter)).to.be.false
+    })
+    test('should pass filter with valid multi transactions', async () => {
+      const transaction = SWAP_MULTI
+      const filter = await swap({
+        chainId: ARB_ONE_CHAIN_ID,
+        contractAddress: AUGUSTUS_SWAPPER_ARBITRUM,
+        tokenIn: Tokens.USDCE,
+        tokenOut: Tokens.VELA,
+        amountOut: GreaterThanOrEqual(parseEther('0.037')),
+      })
+      expect(apply(transaction, filter)).to.be.true
+    })
+    test('should pass filter with valid balancer transaction', async () => {
+      const transaction = SWAP_BALANCER
+      const filter = await swap({
+        chainId: ARB_ONE_CHAIN_ID,
+        contractAddress: AUGUSTUS_SWAPPER_ARBITRUM,
+        tokenIn: Tokens.ETH,
+        tokenOut: '0xEC70Dcb4A1EFa46b8F2D97C310C9c4790ba5ffA8',
+        amountOut: GreaterThanOrEqual(parseEther('0.037')),
+      })
+      expect(apply(transaction, filter)).to.be.true
+    })
+    test('should pass filter with valid curve transaction', async () => {
+      const transaction = SWAP_CURVE
+      const filter = await swap({
+        chainId: ARB_ONE_CHAIN_ID,
+        contractAddress: AUGUSTUS_SWAPPER_ARBITRUM,
+        tokenIn: Tokens.ETH,
+        tokenOut: Tokens.USDT,
+        amountIn: GreaterThanOrEqual(parseEther('0.121')),
+        amountOut: GreaterThanOrEqual(parseUnits('249.50', 6)),
+      })
+      expect(apply(transaction, filter)).to.be.true
+    })
+    test('should pass filter with valid megaswap transaction', async () => {
+      const transaction = SWAP_MEGA
+      const filter = await swap({
+        chainId: ARB_ONE_CHAIN_ID,
+        tokenIn: '0xB64E280e9D1B5DbEc4AcceDb2257A87b400DB149', // LVL Token
+        tokenOut: Tokens.USDT,
+        amountIn: GreaterThanOrEqual(parseUnits('500', 18)),
+        amountOut: GreaterThanOrEqual(parseUnits('100', 6)),
+      })
+      expect(apply(transaction, filter)).to.be.true
+    })
+    test('should not pass filter with invalid transactions', async () => {
+      const transaction = SWAP_MULTI
+      const filter = await swap({
+        chainId: ARB_ONE_CHAIN_ID,
+        contractAddress: AUGUSTUS_SWAPPER_ARBITRUM,
+        tokenIn: Tokens.USDT,
+        tokenOut: Tokens.VELA,
+        amountOut: GreaterThanOrEqual(parseEther('0.037')),
+      })
+      expect(apply(transaction, filter)).to.be.false
+    })
+
+    describe('when returning supported tokens (swap)', () => {
+      const chainIds = [1, 10, 42161]
+      for (const chainId of chainIds) {
+        test(`should return list of tokens for chainId ${chainId}`, async () => {
+          const tokens = await getSupportedTokenAddresses(
+            chainId,
+            ActionType.Swap,
+          )
+          expect(tokens).to.not.be.empty
+        })
+      }
+    })
+  })
+  describe('When handling the stake', () => {
+    test('should pass with a simple deposit', async () => {
+      const transaction = SIMPLE_DEPOSIT
+      const filter = await stake({
+        chainId: OPTIMISM_CHAIN_ID,
+        contractAddress: '0x8C934b7dBc782568d14ceaBbEAeDF37cB6348615',
+        amountOne: GreaterThanOrEqual(parseEther('46128')),
+      })
+      expect(apply(transaction, filter)).to.be.true
+    })
+    test('should pass with a multi deposit', async () => {
+      const transaction = MULTI_DEPOSIT
+      const filter = await stake({
+        chainId: OPTIMISM_CHAIN_ID,
+        contractAddress: AUGUSTUS_SWAPPER_ARBITRUM,
+        amountOne: GreaterThanOrEqual(parseEther('659')),
+      })
+      expect(apply(transaction, filter)).to.be.true
     })
   })
 })