+++ packages/symbiosis/src/index.ts
@@ -1,0 +1,19 @@
+import {
+  type IActionPlugin,
+  PluginActionNotImplementedError,
+} from '@rabbitholegg/questdk'
+
+import {
+  bridge,
+  getSupportedChainIds,
+  getSupportedTokenAddresses,
+} from './Symbiosis.js'
+
+export const Symbiosis: IActionPlugin = {
+  pluginId: 'symbiosis',
+  getSupportedTokenAddresses,
+  getSupportedChainIds,
+  bridge,
+  swap: async () => new PluginActionNotImplementedError(),
+  mint: async () => new PluginActionNotImplementedError(),
+}